refactor: form-validation types
