--- src/lib/form-validation/common/type.ts
@@ -1,12 +1,0 @@
-import { CurrencyExt } from '@interlay/interbtc-api';
-import { MonetaryAmount } from '@interlay/monetary-js';
-
-type CommonValidationParams = {
-  governanceBalance: MonetaryAmount<CurrencyExt>;
-  transactionFee: MonetaryAmount<CurrencyExt>;
-  availableBalance: MonetaryAmount<CurrencyExt>;
-  minAmount: MonetaryAmount<CurrencyExt>;
-  maxAmount?: MonetaryAmount<CurrencyExt>;
-};
-
-export type { CommonValidationParams };
